Sticker factory added
